Added Animation to Name on Hover
