+++ src/app/globals.css
@@ -14,6 +14,13 @@
   }
 }
 
+body {
+  color: var(--foreground);
+  background: var(--background);
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+/* Start of Scrollbar */
 html:not(:root) {
   scrollbar-width: thin;
   scrollbar-color: #475569 #0f172a; /* Thumb: slate-600, Track: slate-900 */
@@ -38,13 +45,9 @@ html:not(:root) {
 ::-webkit-scrollbar-thumb:active {
   @apply bg-slate-400;
 }
+/* End of Scrollbar */
 
-body {
-  color: var(--foreground);
-  background: var(--background);
-  font-family: Arial, Helvetica, sans-serif;
-}
-
+/* Start of Navbar */
 .nav-indicator {
   @apply mr-4 h-px w-8 bg-slate-600 transition-all group-hover:w-16 group-hover:bg-slate-200 group-focus-visible:w-16 group-focus-visible:bg-slate-200 motion-reduce:transition-none;
 }
@@ -60,3 +63,135 @@ body {
 .active .nav-text {
   @apply text-slate-200;
 }
+/* End of Navbar */
+
+/* Start of Animated Name */
+@keyframes float {
+  from, to {
+    transform: translateY(-0%);
+  }
+
+  50% {
+    transform: translateY(-3%);
+  }
+}
+
+@keyframes background-pan {
+  from {
+    background-position: 1000% center;
+  }
+
+  to {
+    background-position: 0% center;
+  }
+}
+
+#name {
+  transition: transform 350ms ease;
+}
+
+#name:hover {
+  transform: scale(1.1);
+  transform-origin: center;
+  font-size: 1em;
+  box-shadow: 0 0 100px rgba(0, 0, 0, 0.75);
+    border-radius: 9999px;
+}
+
+#name:hover::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  border-radius: inherit;
+  box-shadow: inset 0 0 100px rgba(0, 0, 0, 0.25);
+  pointer-events: none;
+}
+
+@media (min-width: 768px) {
+  #name:hover {
+    transform: scale(1.2);
+  }
+}
+
+@media (min-width: 1440px) {
+  #name:hover {
+    transform: scale(1.5);
+  }
+}
+
+#name span {
+  display: inline-block;
+}
+
+#name > .outer {
+  transition:transform 350ms ease;
+}
+
+#name:hover > .outer {
+  transition-duration: 800ms;
+}
+
+#name:hover > .outer > .inner {
+  animation: float 3s ease infinite;
+}
+
+#name:hover > .outer > .inner > .letter {
+  background: linear-gradient(
+    to right,
+    #3b82f6,
+    #22c55e,
+    #3b82f6
+  );
+  background-size: 1000%;
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  animation: background-pan 150s linear infinite;
+}
+
+#name:hover > span:nth-child(1) {
+  transform: translate(-80%, 60%) rotate(8deg);
+}
+#name:hover > span:nth-child(2) {
+  transform: translate(-40%, 20%) rotate(4deg);
+}
+
+#name:hover > span:nth-child(3) {
+  transform: translate(-10%, 60%) rotate(-6deg);
+}
+
+#name:hover > span:nth-child(4) {
+  transform: translate(0%, 8%) rotate(-8deg);
+}
+
+#name:hover > span:nth-child(5) {
+  transform: translate(0%, -20%) rotate(5deg);
+}
+
+#name:hover > span:nth-child(6) {
+  transform: translate(0%, 20%) rotate(-3deg);
+}
+
+#name:hover > span:nth-child(7) {
+  transform: translate(0%, -40%) rotate(-5deg);
+}
+
+#name:hover > span:nth-child(8) {
+  transform: translate(0%, 15%) rotate(10deg);
+}
+
+#name:hover > span:nth-child(9) {
+  transform: translate(0%, -50%) rotate(8deg);
+}
+
+#name:hover > span:nth-child(10) {
+  transform: translate(0%, 15%) rotate(-6deg);
+}
+
+#name:hover > span:nth-child(11) {
+  transform: translate(50%, -10%) rotate(-3deg);
+}
+/* End of Animated Name */
